Modify inbox filters to be moved into filter trait
